requirements.txt
Django==2.2.1
djangorestframework==3.9.3
djangorestframework-stubs==0.4.2
mysqlclient==1.4.2.post1
pytz==2019.1
sqlparse==0.3.0
